+++ packages/core/client/src/providers/index.tsx
@@ -1,1 +1,3 @@
 export * from './ConfigProvider'
+export * from './PubSubProvider'
+export * from './FeathersProvider'
