+++ packages/core/shared/src/types.ts
@@ -239,7 +239,7 @@ export type PubSubEvents = {
   $CREATE_PROJECT_WINDOW: (tabId: string) => string
   $CREATE_DEBUG_CONSOLE: (tabId: string) => string
   $CREATE_CONSOLE: (tabId: string) => string
-  $RUN_SPELL: (tabId: string) => string
+  $RUN_SPELL: (tabId?: string) => string
   $PROCESS: (tabId: string) => string
   $EXPORT: (tabId: string) => string
   $UNDO: (tabId: string) => string
