+++ packages/editor/src/components/MenuBar/MenuBar.tsx
@@ -5,7 +5,7 @@ import { useHotkeys } from 'react-hotkeys-hook'
 import { useDispatch, useSelector } from 'react-redux'
 import { useNavigate } from 'react-router-dom'
 import { useModal } from '../../contexts/ModalProvider'
-import { usePubSub } from '../../contexts/PubSubProvider'
+import { usePubSub } from '../../../../core/client/src/providers/PubSubProvider'
 import { toggleAutoSave } from '../../state/preferences'
 import { RootState } from '../../state/store'
 import { activeTabSelector, changeEditorLayout, Tab } from '../../state/tabs'
