+++ packages/editor/src/contexts/ChatProvider.tsx
@@ -9,13 +9,12 @@ import React, {
 } from 'react'
 import { Component } from 'rete/types/engine'
 
+import { useConfig, usePubSub } from '@magickml/client-core'
 import { MyNode } from '../components/Node/Node'
-import { useConfig } from '@magickml/client-core'
-import { useFeathers } from './FeathersProvider'
-import { usePubSub } from './PubSubProvider'
 import { initEditor } from '../editor'
 import { zoomAt } from '../plugins/areaPlugin/zoom-at'
 import { spellApi } from '../state/api/spells'
+import { useFeathers } from '../../../core/client/src/providers/FeathersProvider'
 import { useMagickInterface } from './MagickInterfaceProvider'
 import styles from './styles.module.scss'
