+++ packages/editor/src/contexts/MagickInterfaceProvider.tsx
@@ -1,4 +1,5 @@
 // DOCUMENTED 
+import { useConfig, usePubSub } from '@magickml/client-core';
 import {
   EditorContext,
   GetSpell,
@@ -14,8 +15,6 @@ import {
   SupportedLanguages,
 } from '@magickml/core';
 import { createContext, useContext, useEffect, useRef } from 'react';
-import { useConfig } from '@magickml/client-core';
-import { usePubSub } from '../contexts/PubSubProvider';
 import { spellApi } from '../state/api/spells';
 
 // Create context for EditorContext type
