+++ packages/editor/src/index.ts
@@ -1,11 +1,3 @@
-// DOCUMENTED 
-/**
- * This exports the AppConfig interface from the './contexts/ConfigProvider' file.
- */
-export type {
-  AppConfig
-} from '../../core/client/src/providers/ConfigProvider';
-
 /**
  * This exports the MagickIDE class from './main' file.
  */
