+++ packages/editor/src/screens/Chat/components/EventHandler.tsx
@@ -7,7 +7,7 @@ import md5 from 'md5'
 import {
   useLazyGetSpellByIdQuery
 } from '../../../state/api/spells'
-import { useFeathers } from '../../../contexts/FeathersProvider'
+import { useFeathers } from '../../../../../core/client/src/providers/FeathersProvider'
 
 import { useConfig } from '@magickml/client-core'
